requirements.txt
openai-whisper
transformers
ffmpeg-python
sentencepiece
numpy==1.26.4